app+board-ui+bot+net+tools: egui pass lifecycle adaptation, phase-state UI rework; annotation re-homes
- egui multi-pass API: explicit end_pass() in tests/harnesses, panel
  blocker registration adjustments; phase-state UI reworked (ui_phase_
  state, ui_plugin, timeline, phase_banner).
- picker/river_placement/zoc/melee/retreat/fire tracks follow the rules
  engine's stacking and FoK cap changes; bot adapts to the same APIs.
- replay_reliability harness: parameter retry-fix and settle knobs.
- app-side traceability annotations re-homed: scenario_maps_to_board +
  start_game_scenario_selects_board to §9.31, campaign_has_no_fixed_
  placements to §9.111.
- traceability-typst main.rs: drop obsolete imports.

+++ omdurman-bot/src/playthrough.rs
@@ -62,7 +62,7 @@ impl KeepOutZone {
             }
             _ => None,
         };
-        ends.is_some_and(|to| to.distance(self.center) as u32 <= self.radius)
+        ends.is_some_and(|to| to.distance(self.center) <= self.radius)
     }
 }
 
@@ -167,7 +167,7 @@ pub async fn playthrough(
         // empty check so a fully-blocked phase falls through to the
         // AdvancePhase escape below.
         if let Some(kz) = cfg.keep_out
-            && kz.in_force(state.active_player, state.current_turn.value() as u8)
+            && kz.in_force(state.active_player, state.current_turn.value())
         {
             candidates.retain(|c| !kz.forbids(c));
         }
